pyproject.toml
[project]
name = "easy-automation-mcp"
version = "0.1.2"
description = "Easy automation MCP server powered by PyAutoGUI — mouse, keyboard, screenshot, and YAML-based flow execution"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "PNL-toshiyaishihara" }]
keywords = ["mcp", "automation", "gui", "pyautogui", "mouse", "keyboard", "screenshot", "flow"]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Desktop Environment",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "fastmcp>=3.2.0",
    "pillow>=12.2.0",
    "pyautogui>=0.9.54",
    "keyboard>=0.13.5",
    "pyyaml>=6.0",
    "mouse>=0.7.1",
]

[project.urls]
Homepage = "https://github.com/PNL-toshiyaishihara/EasyAutomationMCP"
Repository = "https://github.com/PNL-toshiyaishihara/EasyAutomationMCP"
"Bug Tracker" = "https://github.com/PNL-toshiyaishihara/EasyAutomationMCP/issues"

[project.scripts]
easy-automation-mcp = "easy_automation_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "ruff>=0.8.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/easy_automation_mcp"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
